requirements.txt
Django==1.4.1
PIL==1.1.7
South==0.7.6
boto==2.6.0
distribute==0.6.34
dj-database-url==0.2.0
django-dajax==0.9.2
django-dajaxice==0.5.2
git+https://github.com/arneb/django-messages.git#egg=django-messages
django-model-utils==1.1.0
django-registration==0.8
django-storages==1.1.5
gunicorn==0.16.1
psycopg2==2.4.5
pytz==2012h
requests==0.14.2
sorl-thumbnail==11.12
stripe==1.7.6
wsgiref==0.1.2
newrelic
django-sslify>=0.2
